Moved MockLog to lib/mock
