+++ packages/core/src/lib/mock/MockLog.ts
